+++ shoop/core/signals.py
@@ -11,3 +11,4 @@ from django.dispatch import Signal
 get_visibility_errors = Signal(providing_args=["shop_product", "customer"], use_caching=True)
 get_orderability_errors = Signal(providing_args=["shop_product", "customer", "supplier", "quantity"], use_caching=True)
 shipment_created = Signal(providing_args=["order", "shipment"], use_caching=True)
+refund_created = Signal(providing_args=["order", "refund_lines"], use_caching=True)